quantize: compact groups struct, short-block qs stack & inv_width_lut (#1487)
- Compact CoderInfo.groups fields n and len to uint8_t in libfaac/coder.h,
  reducing struct size by 28 bytes per channel instance.
- Optimize assign_band_codebooks() in libfaac/quantize.c to use a 1,024-byte
  local stack buffer short_qs_buf for short blocks, avoiding dynamic heap
  allocations on transient frames.
- Precompute inv_width_lut[128] during QuantizeInit() in libfaac/quantize.c,
  replacing floating-point division in band RMS calculations with multiplication.
- Net .text binary size reduction across optimization passes: -326 bytes.

+++ libfaac/coder.h
@@ -88,8 +88,8 @@ typedef struct CoderInfo {
     const int *sfb_offset;
 
     struct {
-        int n;
-        int len[MAX_SHORT_WINDOWS];
+        uint8_t n;
+        uint8_t len[MAX_SHORT_WINDOWS];
     } groups;
 
     /* worst case: one codeword with two escapes per two spectral lines */
